Crate rusoto_apigatewayv2

Amazon API Gateway V2

If you're using the service, you're probably looking for ApiGatewayV2Client and ApiGatewayV2.

Structs

AccessLogSettings
Settings for logging access in a stage.

Api
Represents an API.

ApiGatewayV2Client
A client for the AmazonApiGatewayV2 API.

ApiMapping
Represents an API mapping.

Apis
Represents a collection of APIs.

Authorizer
Represents an authorizer.

Authorizers
Represents a collection of authorizers.

CreateApiInput
Represents the input parameters for a CreateApi request.

CreateApiMappingInput
Represents the input parameters for a CreateApiMapping request.

CreateApiMappingRequest

CreateApiMappingResponse

CreateApiRequest

CreateApiResponse

CreateAuthorizerInput
Represents the input parameters for a CreateAuthorizer request.

CreateAuthorizerRequest

CreateAuthorizerResponse

CreateDeploymentInput
Represents the input parameters for a CreateDeployment request.

CreateDeploymentRequest

CreateDeploymentResponse

CreateDomainNameInput
Represents the input parameters for a CreateDomainName request.

CreateDomainNameRequest

CreateDomainNameResponse

CreateIntegrationInput
Represents the input parameters for a CreateIntegration request.

CreateIntegrationRequest

CreateIntegrationResponseInput
Represents the input parameters for a CreateIntegrationResponse request.

CreateIntegrationResponseRequest

CreateIntegrationResponseResponse

CreateIntegrationResult

CreateModelInput
Represents the input parameters for a CreateModel request.

CreateModelRequest

CreateModelResponse

CreateRouteInput
Represents the input parameters for a CreateRoute request.

CreateRouteRequest

CreateRouteResponseInput
Represents the input parameters for an CreateRouteResponse request.

CreateRouteResponseRequest

CreateRouteResponseResponse

CreateRouteResult

CreateStageInput
Represents the input parameters for a CreateStage request.

CreateStageRequest

CreateStageResponse

DeleteApiMappingRequest

DeleteApiRequest

DeleteAuthorizerRequest

DeleteDeploymentRequest

DeleteDomainNameRequest

DeleteIntegrationRequest

DeleteIntegrationResponseRequest

DeleteModelRequest

DeleteRouteRequest

DeleteRouteResponseRequest

DeleteStageRequest

Deployment
An immutable representation of an API that can be called by users. A Deployment must be associated with a Stage for it to be callable over the internet.

Deployments
A collection resource that contains zero or more references to your existing deployments, and links that guide you on how to interact with your collection. The collection offers a paginated view of the contained deployments.

DomainName
Represents a domain name.

DomainNameConfiguration
The domain name configuration.

DomainNames
Represents a collection of domain names.

GetApiMappingRequest

GetApiMappingResponse

GetApiMappingsRequest

GetApiMappingsResponse

GetApiRequest

GetApiResponse

GetApisRequest

GetApisResponse

GetAuthorizerRequest

GetAuthorizerResponse

GetAuthorizersRequest

GetAuthorizersResponse

GetDeploymentRequest

GetDeploymentResponse

GetDeploymentsRequest

GetDeploymentsResponse

GetDomainNameRequest

GetDomainNameResponse

GetDomainNamesRequest

GetDomainNamesResponse

GetIntegrationRequest

GetIntegrationResponseRequest

GetIntegrationResponseResponse

GetIntegrationResponsesRequest

GetIntegrationResponsesResponse

GetIntegrationResult

GetIntegrationsRequest

GetIntegrationsResponse

GetModelRequest

GetModelResponse

GetModelTemplateRequest

GetModelTemplateResponse

GetModelsRequest

GetModelsResponse

GetRouteRequest

GetRouteResponseRequest

GetRouteResponseResponse

GetRouteResponsesRequest

GetRouteResponsesResponse

GetRouteResult

GetRoutesRequest

GetRoutesResponse

GetStageRequest

GetStageResponse

GetStagesRequest

GetStagesResponse

Integration
Represents an integration.

IntegrationResponse
Represents an integration response.

IntegrationResponses
Represents a collection of integration responses.

Integrations
Represents a collection of integrations.

LimitExceededException
A limit has been exceeded. See the accompanying error message for details.

Model
Represents a data model for an API. See Create Models and Mapping Templates for Request and Response Mappings.

Models
Represents a collection of data models. See Create Models and Mapping Templates for Request and Response Mappings.

ParameterConstraints
Validation constraints imposed on parameters of a request (path, query string, headers).

Route
Represents a route.

RouteResponse
Represents a route response.

RouteResponses
Represents a collection of route responses.

RouteSettings
Represents a collection of route settings.

Routes
Represents a collection of routes.

Stage
Represents an API stage.

Stages
A collection of Stage resources that are associated with the ApiKey resource.

Template
Represents a template.

UpdateApiInput
Represents the input parameters for an UpdateApi request.

UpdateApiMappingInput
Represents the input parameters for an UpdateApiMapping request.

UpdateApiMappingRequest

UpdateApiMappingResponse

UpdateApiRequest

UpdateApiResponse

UpdateAuthorizerInput
The input parameters for an UpdateAuthorizer request.

UpdateAuthorizerRequest

UpdateAuthorizerResponse

UpdateDeploymentInput
Represents the input parameters for an UpdateDeployment request.

UpdateDeploymentRequest

UpdateDeploymentResponse

UpdateDomainNameInput
Represents the input parameters for an UpdateDomainName request.

UpdateDomainNameRequest

UpdateDomainNameResponse

UpdateIntegrationInput
Represents the input parameters for an UpdateIntegration request.

UpdateIntegrationRequest

UpdateIntegrationResponseInput
Represents the input parameters for an UpdateIntegrationResponse request.

UpdateIntegrationResponseRequest

UpdateIntegrationResponseResponse

UpdateIntegrationResult

UpdateModelInput
Represents the input parameters for an UpdateModel request.

UpdateModelRequest

UpdateModelResponse

UpdateRouteInput
Represents the input parameters for an UpdateRoute request.

UpdateRouteRequest

UpdateRouteResponseInput
Represents the input parameters for an UpdateRouteResponse request.

UpdateRouteResponseRequest

UpdateRouteResponseResponse

UpdateRouteResult

UpdateStageInput
Represents the input parameters for an UpdateStage request.

UpdateStageRequest

UpdateStageResponse

Enums

CreateApiError
Errors returned by CreateApi

CreateApiMappingError
Errors returned by CreateApiMapping

CreateAuthorizerError
Errors returned by CreateAuthorizer

CreateDeploymentError
Errors returned by CreateDeployment

CreateDomainNameError
Errors returned by CreateDomainName

CreateIntegrationError
Errors returned by CreateIntegration

CreateIntegrationResponseError
Errors returned by CreateIntegrationResponse

CreateModelError
Errors returned by CreateModel

CreateRouteError
Errors returned by CreateRoute

CreateRouteResponseError
Errors returned by CreateRouteResponse

CreateStageError
Errors returned by CreateStage

DeleteApiError
Errors returned by DeleteApi

DeleteApiMappingError
Errors returned by DeleteApiMapping

DeleteAuthorizerError
Errors returned by DeleteAuthorizer

DeleteDeploymentError
Errors returned by DeleteDeployment

DeleteDomainNameError
Errors returned by DeleteDomainName

DeleteIntegrationError
Errors returned by DeleteIntegration

DeleteIntegrationResponseError
Errors returned by DeleteIntegrationResponse

DeleteModelError
Errors returned by DeleteModel

DeleteRouteError
Errors returned by DeleteRoute

DeleteRouteResponseError
Errors returned by DeleteRouteResponse

DeleteStageError
Errors returned by DeleteStage

GetApiError
Errors returned by GetApi

GetApiMappingError
Errors returned by GetApiMapping

GetApiMappingsError
Errors returned by GetApiMappings

GetApisError
Errors returned by GetApis

GetAuthorizerError
Errors returned by GetAuthorizer

GetAuthorizersError
Errors returned by GetAuthorizers

GetDeploymentError
Errors returned by GetDeployment

GetDeploymentsError
Errors returned by GetDeployments

GetDomainNameError
Errors returned by GetDomainName

GetDomainNamesError
Errors returned by GetDomainNames

GetIntegrationError
Errors returned by GetIntegration

GetIntegrationResponseError
Errors returned by GetIntegrationResponse

GetIntegrationResponsesError
Errors returned by GetIntegrationResponses

GetIntegrationsError
Errors returned by GetIntegrations

GetModelError
Errors returned by GetModel

GetModelTemplateError
Errors returned by GetModelTemplate

GetModelsError
Errors returned by GetModels

GetRouteError
Errors returned by GetRoute

GetRouteResponseError
Errors returned by GetRouteResponse

GetRouteResponsesError
Errors returned by GetRouteResponses

GetRoutesError
Errors returned by GetRoutes

GetStageError
Errors returned by GetStage

GetStagesError
Errors returned by GetStages

UpdateApiError
Errors returned by UpdateApi

UpdateApiMappingError
Errors returned by UpdateApiMapping

UpdateAuthorizerError
Errors returned by UpdateAuthorizer

UpdateDeploymentError
Errors returned by UpdateDeployment

UpdateDomainNameError
Errors returned by UpdateDomainName

UpdateIntegrationError
Errors returned by UpdateIntegration

UpdateIntegrationResponseError
Errors returned by UpdateIntegrationResponse

UpdateModelError
Errors returned by UpdateModel

UpdateRouteError
Errors returned by UpdateRoute

UpdateRouteResponseError
Errors returned by UpdateRouteResponse

UpdateStageError
Errors returned by UpdateStage

Traits

ApiGatewayV2
Trait representing the capabilities of the AmazonApiGatewayV2 API. AmazonApiGatewayV2 clients implement this trait.