Struct rusoto_apigatewayv2::ApiGatewayV2Client

pub struct ApiGatewayV2Client { /* fields omitted */ }

A client for the AmazonApiGatewayV2 API.

Methods

impl ApiGatewayV2Client

pub fn new(region: Region) -> ApiGatewayV2Client

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ApiGatewayV2Client where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,