Crate rusoto_apigatewayv2

Amazon API Gateway V2

If you're using the service, you're probably looking for ApiGatewayV2Client and ApiGatewayV2.

Structs

AccessLogSettings
Settings for logging access in a stage.

Api
Represents an API.

ApiGatewayV2Client
A client for the AmazonApiGatewayV2 API.

ApiMapping
Represents an API mapping.

Authorizer
Represents an authorizer.

CreateApiMappingRequest

CreateApiMappingResponse

CreateApiRequest

CreateApiResponse

CreateAuthorizerRequest

CreateAuthorizerResponse

CreateDeploymentRequest

CreateDeploymentResponse

CreateDomainNameRequest

CreateDomainNameResponse

CreateIntegrationRequest

CreateIntegrationResponse

CreateIntegrationResponseRequest

CreateIntegrationResponseResponse

CreateModelRequest

CreateModelResponse

CreateRouteRequest

CreateRouteResponse

CreateRouteResponseRequest

CreateRouteResponseResponse

CreateStageRequest

CreateStageResponse

DeleteApiMappingRequest

DeleteApiRequest

DeleteAuthorizerRequest

DeleteDeploymentRequest

DeleteDomainNameRequest

DeleteIntegrationRequest

DeleteIntegrationResponseRequest

DeleteModelRequest

DeleteRouteRequest

DeleteRouteResponseRequest

DeleteStageRequest

Deployment
An immutable representation of an API that can be called by users. A Deployment must be associated with a Stage for it to be callable over the internet.

DomainName
Represents a domain name.

DomainNameConfiguration
The domain name configuration.

GetApiMappingRequest

GetApiMappingResponse

GetApiMappingsRequest

GetApiMappingsResponse

GetApiRequest

GetApiResponse

GetApisRequest

GetApisResponse

GetAuthorizerRequest

GetAuthorizerResponse

GetAuthorizersRequest

GetAuthorizersResponse

GetDeploymentRequest

GetDeploymentResponse

GetDeploymentsRequest

GetDeploymentsResponse

GetDomainNameRequest

GetDomainNameResponse

GetDomainNamesRequest

GetDomainNamesResponse

GetIntegrationRequest

GetIntegrationResponse

GetIntegrationResponseRequest

GetIntegrationResponseResponse

GetIntegrationResponsesRequest

GetIntegrationResponsesResponse

GetIntegrationsRequest

GetIntegrationsResponse

GetModelRequest

GetModelResponse

GetModelTemplateRequest

GetModelTemplateResponse

GetModelsRequest

GetModelsResponse

GetRouteRequest

GetRouteResponse

GetRouteResponseRequest

GetRouteResponseResponse

GetRouteResponsesRequest

GetRouteResponsesResponse

GetRoutesRequest

GetRoutesResponse

GetStageRequest

GetStageResponse

GetStagesRequest

GetStagesResponse

Integration
Represents an integration.

IntegrationResponse
Represents an integration response.

Model
Represents a data model for an API. See Create Models and Mapping Templates for Request and Response Mappings.

ParameterConstraints
Validation constraints imposed on parameters of a request (path, query string, headers).

Route
Represents a route.

RouteResponse
Represents a route response.

RouteSettings
Represents a collection of route settings.

Stage
Represents an API stage.

UpdateApiMappingRequest

UpdateApiMappingResponse

UpdateApiRequest

UpdateApiResponse

UpdateAuthorizerRequest

UpdateAuthorizerResponse

UpdateDeploymentRequest

UpdateDeploymentResponse

UpdateDomainNameRequest

UpdateDomainNameResponse

UpdateIntegrationRequest

UpdateIntegrationResponse

UpdateIntegrationResponseRequest

UpdateIntegrationResponseResponse

UpdateModelRequest

UpdateModelResponse

UpdateRouteRequest

UpdateRouteResponse

UpdateRouteResponseRequest

UpdateRouteResponseResponse

UpdateStageRequest

UpdateStageResponse

Enums

CreateApiError
Errors returned by CreateApi

CreateApiMappingError
Errors returned by CreateApiMapping

CreateAuthorizerError
Errors returned by CreateAuthorizer

CreateDeploymentError
Errors returned by CreateDeployment

CreateDomainNameError
Errors returned by CreateDomainName

CreateIntegrationError
Errors returned by CreateIntegration

CreateIntegrationResponseError
Errors returned by CreateIntegrationResponse

CreateModelError
Errors returned by CreateModel

CreateRouteError
Errors returned by CreateRoute

CreateRouteResponseError
Errors returned by CreateRouteResponse

CreateStageError
Errors returned by CreateStage

DeleteApiError
Errors returned by DeleteApi

DeleteApiMappingError
Errors returned by DeleteApiMapping

DeleteAuthorizerError
Errors returned by DeleteAuthorizer

DeleteDeploymentError
Errors returned by DeleteDeployment

DeleteDomainNameError
Errors returned by DeleteDomainName

DeleteIntegrationError
Errors returned by DeleteIntegration

DeleteIntegrationResponseError
Errors returned by DeleteIntegrationResponse

DeleteModelError
Errors returned by DeleteModel

DeleteRouteError
Errors returned by DeleteRoute

DeleteRouteResponseError
Errors returned by DeleteRouteResponse

DeleteStageError
Errors returned by DeleteStage

GetApiError
Errors returned by GetApi

GetApiMappingError
Errors returned by GetApiMapping

GetApiMappingsError
Errors returned by GetApiMappings

GetApisError
Errors returned by GetApis

GetAuthorizerError
Errors returned by GetAuthorizer

GetAuthorizersError
Errors returned by GetAuthorizers

GetDeploymentError
Errors returned by GetDeployment

GetDeploymentsError
Errors returned by GetDeployments

GetDomainNameError
Errors returned by GetDomainName

GetDomainNamesError
Errors returned by GetDomainNames

GetIntegrationError
Errors returned by GetIntegration

GetIntegrationResponseError
Errors returned by GetIntegrationResponse

GetIntegrationResponsesError
Errors returned by GetIntegrationResponses

GetIntegrationsError
Errors returned by GetIntegrations

GetModelError
Errors returned by GetModel

GetModelTemplateError
Errors returned by GetModelTemplate

GetModelsError
Errors returned by GetModels

GetRouteError
Errors returned by GetRoute

GetRouteResponseError
Errors returned by GetRouteResponse

GetRouteResponsesError
Errors returned by GetRouteResponses

GetRoutesError
Errors returned by GetRoutes

GetStageError
Errors returned by GetStage

GetStagesError
Errors returned by GetStages

UpdateApiError
Errors returned by UpdateApi

UpdateApiMappingError
Errors returned by UpdateApiMapping

UpdateAuthorizerError
Errors returned by UpdateAuthorizer

UpdateDeploymentError
Errors returned by UpdateDeployment

UpdateDomainNameError
Errors returned by UpdateDomainName

UpdateIntegrationError
Errors returned by UpdateIntegration

UpdateIntegrationResponseError
Errors returned by UpdateIntegrationResponse

UpdateModelError
Errors returned by UpdateModel

UpdateRouteError
Errors returned by UpdateRoute

UpdateRouteResponseError
Errors returned by UpdateRouteResponse

UpdateStageError
Errors returned by UpdateStage

Traits

ApiGatewayV2
Trait representing the capabilities of the AmazonApiGatewayV2 API. AmazonApiGatewayV2 clients implement this trait.